+++ falcosidekick-k8s-operator/src/state.py
@@ -10,6 +10,7 @@ from typing import Optional
 
 import ops
 from charms.loki_k8s.v1.loki_push_api import LokiPushApiConsumer
+from charms.traefik_k8s.v2.ingress import IngressPerAppRequirer
 from pydantic import BaseModel, HttpUrl, ValidationError
 
 from config import CharmConfig, InvalidCharmConfigError
@@ -24,11 +25,15 @@ class CharmState(BaseModel):
     the charm configuration and other sources.
 
     Attributes:
+        enable_tls: Whether TLS is enabled for the Falcosidekick workload.
+        http_endpoint_config: The HTTP endpoint configuration dictionary.
         falcosidekick_listenport: The port on which Falcosidekick listens.
         falcosidekick_loki_endpoint: The URL of the Loki push API endpoint.
         falcosidekick_loki_hostport: The host and port of the Loki push API endpoint.
     """
 
+    enable_tls: bool
+    http_endpoint_config: dict
     falcosidekick_listenport: int
     falcosidekick_loki_endpoint: str
     falcosidekick_loki_hostport: str
@@ -38,6 +43,7 @@ class CharmState(BaseModel):
         cls,
         charm: ops.CharmBase,
         loki_push_api_consumer: LokiPushApiConsumer,
+        ingress_requirer: IngressPerAppRequirer,
     ) -> "CharmState":
         """Create a CharmState from a charm instance.
 
@@ -47,6 +53,7 @@ class CharmState(BaseModel):
         Args:
             charm: The charm instance from which to extract state.
             loki_push_api_consumer: The LokiPushApiConsumer instance to get Loki relation data.
+            ingress_requirer: The IngressPerAppRequirer instance to get ingress relation data.
 
         Returns:
             CharmState: A validated CharmState instance.
@@ -60,12 +67,34 @@ class CharmState(BaseModel):
             _url = _get_loki_ingress_endpoint(loki_push_api_consumer)
             loki_endpoint = _url.path if _url and _url.path else "/loki/api/v1/push"
             loki_hostport = f"{_url.scheme}://{_url.host}:{_url.port}" if _url else ""
+            enable_tls = True
+            http_endpoint_config = {
+                "path": "/",
+                "scheme": "https",
+                "set_ports": True,
+                "hostname": None,
+                "listen_port": charm_config.port,
+            }
+            if ingress_requirer.is_ready():
+                ingress_url = HttpUrl(ingress_requirer.url)
+                enable_tls = False  # TLS is handled by ingress
+                http_endpoint_config.update(
+                    {
+                        "path": ingress_url.path,
+                        "scheme": ingress_url.scheme,
+                        "set_ports": False,
+                        "hostname": ingress_url.host,
+                        "listen_port": ingress_url.port,
+                    }
+                )
         except ValidationError as e:
             logger.error("Configuration validation error: %s", e)
             error_fields = set(itertools.chain.from_iterable(err["loc"] for err in e.errors()))
             error_field_str = " ".join(f"{f}" for f in error_fields)
             raise InvalidCharmConfigError(f"Invalid charm configuration: {error_field_str}") from e
         return cls(
+            enable_tls=enable_tls,
+            http_endpoint_config=http_endpoint_config,
             falcosidekick_listenport=charm_config.port,
             falcosidekick_loki_endpoint=loki_endpoint,
             falcosidekick_loki_hostport=loki_hostport,