+++ falcosidekick-k8s-operator/terraform/outputs.tf
@@ -10,6 +10,7 @@ output "requires" {
   value = {
     send_loki_logs = "send-loki-logs"
     certificates   = "certificates"
+    ingress        = "ingress"
   }
 }